dom/grid/test/chrome/test_grid_state.html

<!doctype html>
<html>
<head>
<meta charset="utf-8">
<script type="text/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css" />
<style>
body {
	margin: 40px;
}
.wrapper {
	display: grid;
	width: 600px;
	grid-gap: 0px;
	grid-template-columns: 50px repeat(auto-fit, 100px);
	background-color: #f00;
}
.box {
	background-color: #444;
	color: #fff;
}
.a {
	grid-column: 3;
}
</style>

<script>
'use strict';

SimpleTest.waitForExplicitFinish();

function runTests() {
	var wrapper = document.getElementById("wrapper");
	var grid = wrapper.getGridFragments()[0];
	
	// test count after removal
	is(grid.cols.tracks.length, 2, "Grid column track array compensates for removed auto-fit columns.");
	
	// test state of tracks
	is(grid.cols.tracks[0].state, "static", "Grid column track 0 is marked as static.");
	is(grid.cols.tracks[1].state, "repeat", "Grid column track 1 is marked as repeat.");
	
	SimpleTest.finish();
}
</script>
</head>
<body onLoad="runTests();">

	<div id="wrapper" class="wrapper">
		<div id="boxA" class="box a">A</div>
	</div>
	
</body>
</html>
